--- 031602130+牛康文/Project1/Project1/Debug/Project1.log
@@ -1,2 +1,0 @@
-﻿  源.cpp
-  Project1.vcxproj -> D:\新建文件夹\PersonProject-C\031602130+牛康文\Project1\Debug\Project1.exe
