+++ 031602130+牛康文/wordcount/Project1/result.txt
@@ -1,0 +1,3 @@
+character: -1
+words: 0
+lines: 0
