+++ 031602130+牛康文/wordcount/UnitTest1/stdafx.h
@@ -1,0 +1,13 @@
+// stdafx.h : 标准系统包含文件的包含文件，
+// 或是经常使用但不常更改的
+// 特定于项目的包含文件
+//
+
+#pragma once
+
+#include "targetver.h"
+
+// 用于 CppUnitTest 的头文件
+#include "CppUnitTest.h"
+
+// TODO: 在此处引用程序需要的其他头文件
