+++ 031602130+牛康文/wordcount/UnitTest1/unittest1.cpp
@@ -1,0 +1,18 @@
+#include "stdafx.h"
+#include "CppUnitTest.h"
+
+using namespace Microsoft::VisualStudio::CppUnitTestFramework;
+
+namespace UnitTest1
+{		
+	TEST_CLASS(UnitTest1)
+	{
+	public:
+		
+		TEST_METHOD(TestMethod1)
+		{
+			// TODO: 在此输入测试代码
+		}
+
+	};
+}
